+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### BC breaks
 
   * Raised minimum required PHP version to 5.5
+  * deprecated `stubbles\lang\SecureString` in favor of `stubbles\lang\Secret`, will be removed in 8.0.0
   * removed `stubbles\peer\createBsdSocket()`, `stubbles\peer\BsdSocket` was already removed in 6.0.0